mcpserver/pyproject.toml
[project]
name = "mcp-server-example"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.112.2",
    "fastmcp>=0.1.0",
    "asyncio>=3.4.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
hello = "mcp_server_example:hello"

[tool.uv]
dev-dependencies = [
    "ruff>=0.6.2",
    "fastapi-cli>=0.0.5",
]